Break firestore.js into multiple files where feasible
