+++ mocks/helpers/buildDocFromHash.js
@@ -1,0 +1,14 @@
+module.exports = function buildDocFromHash(hash = {}) {
+  return {
+    exists: !!hash || false,
+    id: hash.id || 'abc123',
+    ref: hash._ref,
+    data() {
+      const copy = { ...hash };
+      delete copy.id;
+      delete copy._collections;
+      delete copy._ref;
+      return copy;
+    },
+  };
+};
